reorganize folders/files and update imports and fix search bar logic
