+++ src/components/PaymentPage/PaymentPage.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import * as Yup from "yup";
 import { stateData } from "../../mockData";
-import { PaymentFormValues } from "../../model";
+import { PaymentFormValues } from "../../resources/model";
 
 const PaymentPage = () => {
   const initialValues: PaymentFormValues = {
