+++ src/components/TestimonialsPage/TestimonialsPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { getReviews } from "../../apiCalls";
-import type { Review } from "../../model";
+import type { Review } from "../../resources/model";
 import holisticHerbalImage from "../../assets/holistic_herbal.jpg";
 
 const TestimonialsPage: React.FC = () => {
